segment ectypes/server.py
